Bug 1328337 - Fix implicit conversion of glyph size from float to uint32_t in NativeFontResource::CreateScaledFont. r=bobowen, a=jcristau
MozReview-Commit-ID: 9JcrfRYJ3ca

+++ gfx/2d/NativeFontResourceDWrite.cpp
@@ -259,7 +259,7 @@ NativeFontResourceDWrite::Create(uint8_t *aFontData, uint32_t aDataLength,
 }
 
 already_AddRefed<ScaledFont>
-NativeFontResourceDWrite::CreateScaledFont(uint32_t aIndex, uint32_t aGlyphSize)
+NativeFontResourceDWrite::CreateScaledFont(uint32_t aIndex, Float aGlyphSize)
 {
   if (aIndex >= mNumberOfFaces) {
     gfxWarning() << "Font face index is too high for font resource.";
